refactor: move a few things into classes
