+++ src/main/java/com/lothrazar/cyclicmagic/ModCyclic.java
@@ -1,4 +1,5 @@
 package com.lothrazar.cyclicmagic;
+import com.lothrazar.cyclicmagic.creativetab.CreativeTabCyclic;
 import com.lothrazar.cyclicmagic.data.Const;
 import com.lothrazar.cyclicmagic.gui.ForgeGuiHandler;
 import com.lothrazar.cyclicmagic.log.ModLogger;
